vfs:uidshift: add helpers to shift UIDs and GIDs
Add helpers to shift UIDs and GIDs according to the user namespace
of the containing mount namespace.

How mounts should appear is defined by the current containing mount
namespace and its user namespace rules, so follow the same logic and use
that user namespace for UID and GID shifting. This gives the advantage
that only a process with the right privileges or an ancestor can set the
user namespace mappings.

Signed-off-by: Dongsu Park <dongsu@endocode.com>
Signed-off-by: Djalal Harouni <djalal@endocode.com>