security: add the security_task_copy() hook
We need to add this hook if we want to add new stackable Linux Security
Modules. There is no other way if LSMs where stackable from the
beginning then that would be easy, however they share the same security
field of creds and they clash on it. Thus if we use the same security
field to store our context this will not work with others LSMs.

Thus we have to add this security hook to capture the context of the
new created task_struct then we can store our own context that refers to
this new task. Using this method allows to add more new stackable LSMs
that will not clash with the already upstreamed LSMs.

Signed-off-by: Djalal Harouni <tixxdz@gmail.com>